+++ x-pack/platform/plugins/shared/alerting_v2/server/lib/rules_client/build_so_search.test.ts
@@ -5,48 +5,7 @@
  * 2.0.
  */
 
-import { buildSoSearch, RULE_SEARCH_FIELDS } from './build_so_search';
-
-describe('buildSoSearch', () => {
-  it('returns undefined for empty input', () => {
-    expect(buildSoSearch()).toBeUndefined();
-    expect(buildSoSearch('')).toBeUndefined();
-    expect(buildSoSearch('   ')).toBeUndefined();
-  });
-
-  it('suffixes a single token with *', () => {
-    expect(buildSoSearch('prod')).toBe('prod*');
-  });
-
-  it('suffixes each token with * for multi-word input', () => {
-    expect(buildSoSearch('prod alerts')).toBe('prod* alerts*');
-  });
-
-  it('trims leading and trailing whitespace', () => {
-    expect(buildSoSearch('  prod alerts  ')).toBe('prod* alerts*');
-  });
-
-  it('collapses multiple spaces between tokens', () => {
-    expect(buildSoSearch('prod    alerts')).toBe('prod* alerts*');
-  });
-
-  it('escapes simple_query_string operators', () => {
-    expect(buildSoSearch('a--b')).toBe('a\\-\\-b*');
-    expect(buildSoSearch('rule+name')).toBe('rule\\+name*');
-    expect(buildSoSearch('(test)')).toBe('\\(test\\)*');
-    expect(buildSoSearch('"quoted"')).toBe('\\"quoted\\"*');
-    expect(buildSoSearch('wild*card')).toBe('wild\\*card*');
-    expect(buildSoSearch('fuzzy~2')).toBe('fuzzy\\~2*');
-    expect(buildSoSearch('back\\slash')).toBe('back\\\\slash*');
-    expect(buildSoSearch('a|b')).toBe('a\\|b*');
-  });
-
-  it('leaves non-operator special characters as-is', () => {
-    expect(buildSoSearch('prod:alerts')).toBe('prod:alerts*');
-    expect(buildSoSearch('#channel')).toBe('#channel*');
-    expect(buildSoSearch('rule@team')).toBe('rule@team*');
-  });
-});
+import { RULE_SEARCH_FIELDS } from './build_so_search';
 
 describe('RULE_SEARCH_FIELDS', () => {
   it('contains only text-mapped fields', () => {
