+++ x-pack/platform/plugins/shared/alerting_v2/server/lib/rules_client/build_so_search.ts
@@ -5,6 +5,8 @@
  * 2.0.
  */
 
+export { buildSoSearch } from '../build_so_search';
+
 /**
  * Fields searched via the SO client's `search` / `searchFields` params
  * (simple_query_string). Only `text`-mapped fields can be listed here —
@@ -16,28 +18,3 @@
  * sub-field to their mapping and reference it here.
  */
 export const RULE_SEARCH_FIELDS = ['metadata.name', 'metadata.description'];
-
-/**
- * Escapes characters that `simple_query_string` treats as operators so
- * user input is matched literally.  Operators: + | - " * ( ) ~ \
- */
-const SIMPLE_QS_OPERATORS = /([+\-|"*()~\\])/g;
-const escapeToken = (token: string): string => token.replace(SIMPLE_QS_OPERATORS, '\\$1');
-
-/**
- * Builds the `search` string for the SO client's simple_query_string query.
- * Each whitespace-delimited token is escaped and suffixed with `*` for
- * prefix matching.
- */
-export const buildSoSearch = (search?: string): string | undefined => {
-  if (!search?.trim()) {
-    return undefined;
-  }
-
-  return search
-    .trim()
-    .split(/\s+/)
-    .filter(Boolean)
-    .map((token) => `${escapeToken(token)}*`)
-    .join(' ');
-};
